perf: get tool client code
