+++ projects/app/src/pageComponents/app/detail/Edit/ChatAgent/utils.ts
@@ -37,7 +37,7 @@ import {
   getToolConfigStatus,
   validateToolConfiguration
 } from '@fastgpt/global/core/app/formEdit/utils';
-import { getToolPreviewNode } from '@/web/core/app/api/tool';
+import { getClientToolPreviewNode } from '@/web/core/app/api/tool';
 import type { AppFileSelectConfigType } from '@fastgpt/global/core/app/type/config.schema';
 import { DatasetSearchModeEnum } from '@fastgpt/global/core/dataset/constants';
 
@@ -431,7 +431,7 @@ export const loadGeneratedTools = async ({
         }
 
         // 新工具，需要与已配置的 tool 进行 input 合并
-        const tool = await getToolPreviewNode({ appId: toolId, versionId: '' });
+        const tool = await getClientToolPreviewNode({ appId: toolId, versionId: '' });
         // 验证工具配置
         const toolValid = validateToolConfiguration({
           toolTemplate: tool,
